+++ Editor/Drawers/HandleDrawers/SphereHandleCapDrawer/SphereHandleCapAttributeDrawer.cs
@@ -73,9 +73,10 @@ namespace SaintsField.Editor.Drawers.HandleDrawers.SphereHandleCapDrawer
 
         private static void UpdateSphereInfo(SphereInfo sphereInfo)
         {
-            Type fieldType = ReflectUtils.GetElementType(sphereInfo.MemberInfo is FieldInfo fi
+            Type rawType = sphereInfo.MemberInfo is FieldInfo fi
                 ? fi.FieldType
-                : ((PropertyInfo)sphereInfo.MemberInfo).PropertyType);
+                : ((PropertyInfo)sphereInfo.MemberInfo).PropertyType;
+            Type fieldType = sphereInfo.SerializedProperty.isArray? ReflectUtils.GetElementType(rawType): rawType;
             bool filedIsNumber = fieldType == typeof(float) || fieldType == typeof(double) || fieldType == typeof(int) || fieldType == typeof(long) || fieldType == typeof(short);
 
             if(filedIsNumber)